requirements.txt
python-novaclient
oslo.config>=1.4.0
